Реализована альбомная ориентация Изменена работа приложения - теперь в основе фрагменты Удалены все лишние классы
