+++ app/src/main/java/ru/geekbrains/noteapphomework/data/Controller.java
@@ -4,6 +4,5 @@ public interface Controller{
     void openEditNoteFragment(Note note);
     void openEditNoteFragmentLandscape(Note note);
     void pressedOkButtonEditNote();
-    void pressedOkButtonEditNoteLandscape();
 }
