+++ app/src/main/res/layout-land/activity_notes_list.xml
@@ -3,11 +3,12 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/note_style"
+
     android:orientation="horizontal">
 
     <LinearLayout
-        android:id="@+id/left_landscape_container"
+        android:id="@+id/fragment_notes_list_container"
+        android:background="@drawable/note_style"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
@@ -16,6 +17,7 @@
 
     <LinearLayout
         android:id="@+id/right_landscape_container"
+        android:background="@drawable/note_style"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
