+++ app/src/main/res/layout/fragment_edit_note.xml
@@ -4,8 +4,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.EditNoteActivity">
+    android:layout_height="match_parent">
 
         <FrameLayout
             android:layout_width="match_parent"
